Adds username to SubAccounts table to distinguish sub accounts.
All sub accounts are listed in the SubAccounts table but there
is no way to tell which account they are subject to. This commit
adds another field for the parent account number. The parent
account number is obtained by requiring username and password
entry to create the new sub account.